Restructured onMove().
Bounds were being generated centered at items position. This will cause problems when moving to SAT for entities and entity management because SAT Shapes have a movable origin used for translating/moving shapes vertices. Bounds on shapes are generated by the vertices not the origin. The bounds on QuadTreeNodeItems will have to mirror the bounds from the shapes, instead of being arbitrarily generated using a center point and dimensions, regardless of the x and y coordinates.