Added functionality.
New 'contains()' method for testing whether one projection contains another. Projections do not know anything about the axis they are on, so be sure you test projections from the same axis to get useful results.

New 'getMin()' and 'getMax()' methods for obtaining the minimum and maximum of the projection area.