Added basic animation class.
Animation uses a SpriteSheet as a source to draw from. It uses a grid location, grid dimensions, and a frame count to evaluate where to retrieve source images from the sprite sheet to draw as animation frames. Animations must also be given a rate at which to draw frames, in order to draw the correct frames over time. Animations currently only loop with no indication on whether the animation has finished. If a frame is evaluated and the coordinates are not on the spritesheet, it will not issue any warning, or autocorrect itself by wrapping down to the next row of suitable spritesheet grid coordinates. This will be addressed in a future build.